Wrapper script for compactor runner
